Move MappedGraph to Separate File
